Remove cmptag.
This doesn't exist in upstream Pawn. It's not clear what it does, or if
it ever did anything, or if it worked correctly when it did. It's
awkwardly attached to value structs even though it's actually parsing
state. Even more strange is that sometimes it's a tag and sometimes it
turns into a symbol count.